Safari-611.1.10/LayoutTests/js/numeric-escapes-in-string-literals-expected.txt

Test numeric escapes in string literals - https://bugs.webkit.org/show_bug.cgi?id=51724

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) is strictResult
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS eval(stringLiteral) is nonStrictResult
PASS eval(stringLiteral) threw exception SyntaxError: The only valid numeric escape in strict mode is '\0'.
PASS successfullyParsed is true

TEST COMPLETE
